requirements.txt
langchain
faiss-cpu
requests
pypdf
pymupdf
python-dotenv
fitz
langchain-community